+++ src/main/java/org/cloudwarp/quickstack/networking/DumpC2SPacket.java
@@ -1,0 +1,17 @@
+package org.cloudwarp.quickstack.networking;
+
+import net.fabricmc.fabric.api.networking.v1.PacketSender;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.network.ServerPlayNetworkHandler;
+import net.minecraft.server.network.ServerPlayerEntity;
+import org.cloudwarp.quickstack.registry.QuickStackDepositFunctions;
+
+public class DumpC2SPacket {
+	public static void receive(MinecraftServer server, ServerPlayerEntity player, ServerPlayNetworkHandler handler, PacketByteBuf buf, PacketSender responseSender){
+		server.execute(() -> {
+			QuickStackDepositFunctions.findNearbyChests(player, false);
+			return;
+		});
+	}
+}