+++ src/main/java/org/cloudwarp/quickstack/registry/QSEventHandler.java
@@ -1,0 +1,13 @@
+package org.cloudwarp.quickstack.registry;
+
+import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.network.PacketByteBuf;
+import org.cloudwarp.quickstack.QuickStack;
+
+public class QSEventHandler {
+
+	public static void registerEvents(){
+	}
+}