+++ changelog.md
@@ -2,4 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.1.0] - 8/14/2022
+## [0.1.0] - 10/5/2022
+### Added
+- Quick stacking to nearby chests
+- Dumping to nearby chests